Organize and enhance comments
