+++ src/PowerShell/Microsoft.WinGet.Client/Commands/GetSourceCommand.cs
@@ -9,7 +9,7 @@ namespace Microsoft.WinGet.Client.Commands
     using System;
     using System.Management.Automation;
     using Microsoft.Management.Deployment;
-    using Microsoft.WinGet.Client.Helpers;
+    using Microsoft.WinGet.Client.Common;
 
     /// <summary>
     /// Retrieves the list of configured sources.
