+++ src/PowerShell/Microsoft.WinGet.Client/Commands/UninstallPackageCommand.cs
@@ -9,6 +9,7 @@ namespace Microsoft.WinGet.Client.Commands
     using System;
     using System.Management.Automation;
     using Microsoft.Management.Deployment;
+    using Microsoft.WinGet.Client.Common;
     using Microsoft.WinGet.Client.Helpers;
 
     /// <summary>
@@ -20,7 +21,7 @@ namespace Microsoft.WinGet.Client.Commands
         DefaultParameterSetName = Constants.FoundSet,
         SupportsShouldProcess = true)]
     [OutputType(typeof(UninstallResult))]
-    public sealed class UninstallPackageCommand : BaseLifecycleCommand
+    public sealed class UninstallPackageCommand : BasePackageCommand
     {
         /// <summary>
         /// Gets or sets the desired mode for the uninstallation process.
@@ -35,23 +36,21 @@ namespace Microsoft.WinGet.Client.Commands
         {
             this.GetPackageAndExecute(CompositeSearchBehavior.LocalCatalogs, (package, version) =>
             {
-                var options = this.GetUninstallOptions(version);
-                var results = this.UninstallPackage(package, options);
-                this.WriteObject(results);
+                UninstallOptions options = this.GetUninstallOptions(version);
+                UninstallResult result = this.UninstallPackage(package, options);
+                this.WriteObject(result);
             });
         }
 
         private UninstallOptions GetUninstallOptions(PackageVersionId version)
         {
             var options = ComObjectFactory.Value.CreateUninstallOptions();
-
             if (this.Log != null)
             {
                 options.LogOutputPath = this.Log;
             }
 
             options.PackageUninstallMode = this.Mode;
-
             if (version != null)
             {
                 options.PackageVersionId = version;
@@ -65,36 +64,27 @@ namespace Microsoft.WinGet.Client.Commands
             UninstallOptions options)
         {
             string activity = $"Uninstalling '{package.Name}'";
-
             var operation = PackageManager.Value.UninstallPackageAsync(package, options);
-            var adapter = new WriteProgressAdapter(this);
-
+            WriteProgressAdapter adapter = new (this);
             operation.Progress = (context, progress) =>
             {
-                var record = new ProgressRecord(1, activity, progress.State.ToString())
+                adapter.WriteProgress(new ProgressRecord(1, activity, progress.State.ToString())
                 {
                     RecordType = ProgressRecordType.Processing,
-                };
-
-                adapter.WriteProgress(record);
+                });
             };
-
             operation.Completed = (context, status) =>
             {
-                var record = new ProgressRecord(1, activity, status.ToString())
+                adapter.WriteProgress(new ProgressRecord(1, activity, status.ToString())
                 {
                     RecordType = ProgressRecordType.Completed,
-                };
-
-                adapter.WriteProgress(record);
+                });
                 adapter.Completed = true;
             };
-
             Console.CancelKeyPress += (sender, e) =>
             {
                 operation.Cancel();
             };
-
             adapter.Wait();
             return operation.GetResults();
         }
