+++ src/PowerShell/Microsoft.WinGet.Client/Microsoft.WinGet.Client.csproj
@@ -1,9 +1,16 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFrameworks>net5.0-windows10.0.22000.0;net461</TargetFrameworks>
-    <Platforms>ARM64;x64;x86</Platforms>
+    <TargetWindowsVersion>10.0.22000.0</TargetWindowsVersion>
+    <CoreFramework>net5.0-windows$(TargetWindowsVersion)</CoreFramework>
+    <DesktopFramework>net461</DesktopFramework>
+  </PropertyGroup>
+  
+  <PropertyGroup>
     <CopyLocalLockFileAssemblies>true</CopyLocalLockFileAssemblies>
+    <LangVersion>9</LangVersion>
+    <Platforms>x64;x86</Platforms>
+    <TargetFrameworks>$(CoreFramework);$(DesktopFramework)</TargetFrameworks>
   </PropertyGroup>
 
   <ItemGroup>
@@ -30,32 +37,32 @@
     </ProjectReference>
   </ItemGroup>
 
-	<ItemGroup Condition="'$(TargetFramework)' == 'net5.0-windows10.0.22000.0'">
+	<ItemGroup Condition="'$(TargetFramework)' == '$(CoreFramework)'">
 		<PackageReference Include="Microsoft.Windows.CsWinRT" Version="1.6.4" />
 	</ItemGroup>
 
-	<ItemGroup Condition="'$(TargetFramework)' == 'net461'">
+	<ItemGroup Condition="'$(TargetFramework)' == '$(DesktopFramework)'">
     <PackageReference Include="Microsoft.Windows.SDK.Contracts" Version="10.0.22000.196" PrivateAssets="all" />
 	</ItemGroup>
 
 	<ItemGroup>
-	  <Compile Update="Properties\Resources.Designer.cs">
+	  <Compile Update="Properties\Resource.Designer.cs">
 	    <DesignTime>True</DesignTime>
 	    <AutoGen>True</AutoGen>
-	    <DependentUpon>Resources.resx</DependentUpon>
+	    <DependentUpon>Resource.resx</DependentUpon>
 	  </Compile>
 	</ItemGroup>
 
 	<ItemGroup>
-	  <EmbeddedResource Update="Properties\Resources.resx">
+	  <EmbeddedResource Update="Properties\Resource.resx">
 	    <Generator>ResXFileCodeGenerator</Generator>
-	    <LastGenOutput>Resources.Designer.cs</LastGenOutput>
+	    <LastGenOutput>Resource.Designer.cs</LastGenOutput>
 	  </EmbeddedResource>
 	</ItemGroup>
 
-	<PropertyGroup Condition="'$(TargetFramework)' == 'net5.0-windows10.0.22000.0'">
+	<PropertyGroup Condition="'$(TargetFramework)' == '$(CoreFramework)'">
 		<CsWinRTIncludes>Microsoft.Management.Deployment</CsWinRTIncludes>
-		<CsWinRTWindowsMetadata>10.0.22000.0</CsWinRTWindowsMetadata>
+		<CsWinRTWindowsMetadata>$(TargetWindowsVersion)</CsWinRTWindowsMetadata>
     <!-- Ensure Support for Windows 10, Version 1809 -->
     <TargetPlatformMinVersion>10.0.17763.0</TargetPlatformMinVersion>
 	</PropertyGroup>
